+++ main.go
@@ -1,39 +1,36 @@
 package main
 
 import (
+	"fmt"
 	"net/http"
+	"os"
+	"strings"
 
 	"github.com/LeoCBS/httpmiddleware"
+	"github.com/LeoCBS/httpmiddleware-app-example/api"
+	"github.com/LeoCBS/httpmiddleware-app-example/storage"
 	"github.com/sirupsen/logrus"
 )
 
-type user struct {
-	Login string `json:"login"`
-}
-
 func main() {
+	mc, err := storage.NewMongoClient(getMongoURI())
+	if err != nil {
+		panic(err)
+	}
+	a := api.New(mc)
 	l := logrus.New()
 	md := httpmiddleware.New(l)
-	md.POST("/name/:name", createUser)
-	md.GET("/name/:name", getUser)
+	a.AddHandlers(md)
 
 	s := &http.Server{
 		Addr:    ":8080",
 		Handler: md,
 	}
+	l.Info("starting server on port 8080")
 	panic(s.ListenAndServe())
 }
 
-func createUser(w http.ResponseWriter, r *http.Request, ps httpmiddleware.Params) httpmiddleware.Response {
-	return httpmiddleware.Response{
-		StatusCode: http.StatusCreated,
-		Body:       user{Login: ps.ByName("name")},
-	}
-}
-
-func getUser(w http.ResponseWriter, r *http.Request, ps httpmiddleware.Params) httpmiddleware.Response {
-	return httpmiddleware.Response{
-		StatusCode: http.StatusCreated,
-		Body:       user{Login: ps.ByName("name")},
-	}
+func getMongoURI() string {
+	mongoPort := strings.ReplaceAll(os.Getenv("MONGO_PORT"), "tcp://", "")
+	return fmt.Sprintf("mongodb://%s", mongoPort)
 }