+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 (unreleased)
+ - Adding main package (@LeoCBS)
  - Adding api package (@LeoCBS)
  - Adding storage/mongo.go package (@LeoCBS)